refactor: order unit test files within tests folder
