+++ test/unit/lib/services/scalingo-client_test.js
@@ -1,8 +1,8 @@
 const { describe, it } = require('mocha');
-const { sinon } = require('../test-helper');
+const { sinon } = require('../../test-helper');
 const scalingo = require('scalingo');
 
-const ScalingoClient = require('../../../lib/services/scalingo-client');
+const ScalingoClient = require('../../../../lib/services/scalingo-client');
 const { expect } = require('chai');
 
 describe('Scalingo client', () => {
