+++ test/unit/scripts/get-pull-requests-to-release-in-prod_test.js
@@ -1,8 +1,8 @@
 const { expect } = require('chai');
 const sinon = require('sinon');
-const github = require('../lib/services/github');
+const github = require('../../../lib/services/github');
 
-const { displayPullRequest, filterPullRequest, getHeadOfChangelog, orderPr, getLastMEPDate } = require('../scripts/get-pull-requests-to-release-in-prod');
+const { displayPullRequest, filterPullRequest, getHeadOfChangelog, orderPr, getLastMEPDate } = require('../../../scripts/get-pull-requests-to-release-in-prod');
 
 describe('Unit | Script | GET Pull Request to release in Prod', () => {
